+++ com.unity.render-pipelines.core/Runtime/RenderGraph/RenderGraphResourceRegistry.cs
@@ -404,11 +404,9 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
 
                     if (resource.desc.clearBuffer || m_RenderGraphDebug.clearRenderTargetsAtCreation)
                     {
-                        // Commented because string.Format causes garbage
-                        //using (new ProfilingSample(rgContext.cmd, string.Format("RenderGraph: Clear Buffer {0}", resourceDescMoved.desc.name)))
                         bool debugClear = m_RenderGraphDebug.clearRenderTargetsAtCreation && !resource.desc.clearBuffer;
                         var name = debugClear ? "RenderGraph: Clear Buffer (Debug)" : "RenderGraph: Clear Buffer";
-                        using (new ProfilingSample(rgContext.cmd, name))
+                        using (new ProfilingScope(rgContext.cmd, ProfilingSampler.Get(RenderGraphProfileId.RenderGraphClear)))
                         {
                             var clearFlag = resource.desc.depthBufferBits != DepthBits.None ? ClearFlag.Depth : ClearFlag.Color;
                             var clearColor = debugClear ? Color.magenta : resource.desc.clearColor;
@@ -501,7 +499,7 @@ namespace UnityEngine.Experimental.Rendering.RenderGraphModule
                 {
                     if (m_RenderGraphDebug.clearRenderTargetsAtRelease)
                     {
-                        using (new ProfilingSample(rgContext.cmd, "RenderGraph: Clear Buffer (Debug)"))
+                        using (new ProfilingScope(rgContext.cmd, ProfilingSampler.Get(RenderGraphProfileId.RenderGraphClearDebug)))
                         {
                             var clearFlag = resourceDesc.desc.depthBufferBits != DepthBits.None ? ClearFlag.Depth : ClearFlag.Color;
                             CoreUtils.SetRenderTarget(rgContext.cmd, GetTexture(resource), clearFlag, Color.magenta);